changes to support vox loading
