+++ src/core/read-stream.js
@@ -33,6 +33,18 @@ class ReadStream {
         return this.offset - amount;
     }
 
+    readChar() {
+        return String.fromCharCode(this.dataView.getUint8(this.offset++));
+    }
+
+    readChars(numChars) {
+        let result = '';
+        for (let i = 0; i < numChars; ++i) {
+            result += this.readChar();
+        }
+        return result;
+    }
+
     readU8() {
         return this.dataView.getUint8(this.offset++);
     }
