+++ src/index.js
@@ -22,6 +22,7 @@ export { platform } from './core/platform.js';
 export { string } from './core/string.js';
 export { EventHandler } from './core/event-handler.js';
 export { IndexedList } from './core/indexed-list.js';
+export { ReadStream } from './core/read-stream.js';
 export { SortedLoopArray } from './core/sorted-loop-array.js';
 export { Tags } from './core/tags.js';
 export { Timer, now } from './core/time.js';
@@ -98,6 +99,7 @@ export { MorphInstance } from './scene/morph-instance.js';
 export { MorphTarget } from './scene/morph-target.js';
 export { ParticleEmitter } from './scene/particle-system/particle-emitter.js';
 export { Picker } from './scene/picker.js';
+export { Render } from './scene/render.js';
 export { Scene } from './scene/scene.js';
 export { Skin } from './scene/skin.js';
 export { SkinInstance } from './scene/skin-instance.js';
@@ -151,7 +153,8 @@ export { AnimationHandler } from './resources/animation.js';
 export { AudioHandler } from './resources/audio.js';
 export { BinaryHandler } from './resources/binary.js';
 export { BundleHandler } from './resources/bundle.js';
-export { ContainerHandler, ContainerResource } from './resources/container.js';
+export { ContainerHandler } from './resources/container.js';
+export { ContainerResource } from './resources/parser/container-resource.js';
 export { createStyle, CssHandler } from './resources/css.js';
 export { CubemapHandler } from './resources/cubemap.js';
 export { FolderHandler } from './resources/folder.js';
